+++ api/src/main/java/pl/kolendateam/dadcard/characterCard/repository/CharacterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pl.kolendateam.dadcard.characterCard.entity.Character;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<Character, Integer> {
+public interface CharacterRepository extends JpaRepository<Character, Short> {
     
 }
