+++ api/src/main/java/pl/kolendateam/dadcard/classCharacter/dto/ClassPcDTO.java
@@ -5,7 +5,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ClassPcDTO {
 
-    public int id;
+    public short id;
     public String className;
     public int level;
     public int[] classStudyId;
