+++ api/src/main/java/pl/kolendateam/dadcard/classCharacter/repository/ClassRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.kolendateam.dadcard.classCharacter.entity.ClassCharacter;
 
-public interface ClassRepository extends JpaRepository<ClassCharacter, Integer>{
+public interface ClassRepository extends JpaRepository<ClassCharacter, Short>{
 
     Optional <ClassCharacter> findByName(String className);
