+++ api/src/main/java/pl/kolendateam/dadcard/items/WeaponsController.java
@@ -45,7 +45,7 @@ public class WeaponsController {
     }
 
     @PostMapping(value = "{id}", consumes = {"application/json"})
-    public CharacterDTO buyWeapons(@PathVariable int id, @RequestBody WeaponsDTO wDTO){
+    public CharacterDTO buyWeapons(@PathVariable short id, @RequestBody WeaponsDTO wDTO){
         Optional<Character> characterOpt = this.characterRepository.findById(id);
 
         if (!characterOpt.isPresent()) {
@@ -73,7 +73,7 @@ public class WeaponsController {
     }
 
     @PostMapping(value = "{id}/sell", consumes = {"application/json"})
-    public CharacterDTO sellWeapons(@PathVariable int id, @RequestBody WeaponsDTO wDTO){
+    public CharacterDTO sellWeapons(@PathVariable short id, @RequestBody WeaponsDTO wDTO){
         Optional<Character> characterOpt = this.characterRepository.findById(id);
 
         if (!characterOpt.isPresent()) {
